pyproject.toml
[project]
name = "llm-classifiers"
version = "0.1.0"
description = "Framework to build production-grade LLM-based classifiers"
readme = "README.md"
authors = [
    { name = "Mads Møller", email = "mads.moeller@outlook.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.10.6",
    "pydantic-ai>=0.0.41",
    "pydantic-logfire>=0.0.1",
    "pandas>=2.0.0",
]

[dependency-groups]
dev = [
    "flake8>=7.1.2",
    "flake8-docstrings>=1.7.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["llm_classifiers"]
